# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RoadScan - Pothole & Litter Detection API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: detections
    description: Detection operations
  - name: stats
    description: Statistics operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /detections:
    get:
      operationId: listDetections
      tags: [detections]
      summary: List all past detections
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: mediaType
          in: query
          schema:
            type: string
            enum: [image, video, all]
            default: all
      responses:
        "200":
          description: List of detections
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DetectionListResponse"

  /detections/{id}:
    get:
      operationId: getDetection
      tags: [detections]
      summary: Get a single detection by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Detection detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Detection"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDetection
      tags: [detections]
      summary: Delete a detection record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get global detection statistics
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StatsResponse"

  /recent:
    get:
      operationId: getRecentDetections
      tags: [detections]
      summary: Get 5 most recent detections
      responses:
        "200":
          description: Recent detections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Detection"

  /detections/map:
    get:
      operationId: getMapDetections
      tags: [detections]
      summary: Get all detections that have GPS coordinates
      responses:
        "200":
          description: Detections with GPS
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Detection"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string

    BoundingBox:
      type: object
      required: [x, y, width, height]
      properties:
        x:
          type: number
          description: Left edge (0-1 relative)
        y:
          type: number
          description: Top edge (0-1 relative)
        width:
          type: number
          description: Box width (0-1 relative)
        height:
          type: number
          description: Box height (0-1 relative)

    DetectedObject:
      type: object
      required: [id, className, confidence, bbox]
      properties:
        id:
          type: string
        className:
          type: string
          enum: [pothole, plastic_waste, other_litter]
        confidence:
          type: number
          description: Confidence score 0-1
        bbox:
          $ref: "#/components/schemas/BoundingBox"
        frameNumber:
          type: ["integer", "null"]
          description: Frame number for video detections

    Detection:
      type: object
      required: [id, filename, mediaType, originalUrl, annotatedUrl, objects, counts, processingTimeMs, severity, createdAt]
      properties:
        id:
          type: integer
        filename:
          type: string
        mediaType:
          type: string
          enum: [image, video]
        originalUrl:
          type: string
          description: URL to the original uploaded file
        annotatedUrl:
          type: string
          description: URL to the annotated output file
        thumbnailUrl:
          type: ["string", "null"]
          description: URL to a small thumbnail
        objects:
          type: array
          items:
            $ref: "#/components/schemas/DetectedObject"
        counts:
          $ref: "#/components/schemas/DetectionCounts"
        processingTimeMs:
          type: integer
        severity:
          type: string
          enum: [low, medium, high, critical]
          description: Computed severity based on detections
        location:
          type: ["string", "null"]
          description: Optional location label
        lat:
          type: ["number", "null"]
          description: GPS latitude from EXIF or manual input
        lon:
          type: ["number", "null"]
          description: GPS longitude from EXIF or manual input
        reporterName:
          type: ["string", "null"]
          description: Reporter name for public view
        createdAt:
          type: string
          format: date-time

    DetectionCounts:
      type: object
      required: [pothole, plastic_waste, other_litter, total]
      properties:
        pothole:
          type: integer
        plastic_waste:
          type: integer
        other_litter:
          type: integer
        total:
          type: integer

    DetectionListResponse:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Detection"
        total:
          type: integer

    StatsResponse:
      type: object
      required: [totalScans, totalObjects, classBreakdown, severityBreakdown, avgConfidence, avgProcessingMs, recentActivity]
      properties:
        totalScans:
          type: integer
        totalObjects:
          type: integer
        classBreakdown:
          $ref: "#/components/schemas/DetectionCounts"
        severityBreakdown:
          type: object
          required: [low, medium, high, critical]
          properties:
            low:
              type: integer
            medium:
              type: integer
            high:
              type: integer
            critical:
              type: integer
        avgConfidence:
          type: number
        avgProcessingMs:
          type: number
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/DailyActivity"

    DailyActivity:
      type: object
      required: [date, scans, objects]
      properties:
        date:
          type: string
        scans:
          type: integer
        objects:
          type: integer