# lib/api-zod/src/generated/types/boundingBox.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * RoadScan - Pothole & Litter Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface BoundingBox {
  /** Left edge (0-1 relative) */
  x: number;
  /** Top edge (0-1 relative) */
  y: number;
  /** Box width (0-1 relative) */
  width: number;
  /** Box height (0-1 relative) */
  height: number;
}