# lib/api-zod/src/generated/types/detectedObject.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * RoadScan - Pothole & Litter Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { BoundingBox } from './boundingBox';
import type { DetectedObjectClassName } from './detectedObjectClassName';

export interface DetectedObject {
  id: string;
  className: DetectedObjectClassName;
  /** Confidence score 0-1 */
  confidence: number;
  bbox: BoundingBox;
  /**
     * Frame number for video detections
     * @nullable
     */
  frameNumber?: number | null;
}