# lib/api-zod/src/generated/types/detectionCounts.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * RoadScan - Pothole & Litter Detection API
 * OpenAPI spec version: 0.1.0
 */

export interface DetectionCounts {
  pothole: number;
  plastic_waste: number;
  other_litter: number;
  total: number;
}