# lib/api-zod/src/generated/types/statsResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * RoadScan - Pothole & Litter Detection API
 * OpenAPI spec version: 0.1.0
 */
import type { DailyActivity } from './dailyActivity';
import type { DetectionCounts } from './detectionCounts';
import type { StatsResponseSeverityBreakdown } from './statsResponseSeverityBreakdown';

export interface StatsResponse {
  totalScans: number;
  totalObjects: number;
  classBreakdown: DetectionCounts;
  severityBreakdown: StatsResponseSeverityBreakdown;
  avgConfidence: number;
  avgProcessingMs: number;
  recentActivity: DailyActivity[];
}